move token types to lexer header
